Saving notebook versions before change.
Saving this version of the notebook before updating my graph to display
metrics for ALL versions of a given model name.

Main usecase is that a given model is deprecated and replaced by its
newer version. In the graph that will look like one timeseries
disappearing and another taking its place forward in time.