Move dvc repo PAT secret to prefect secrets.
Initially i was storing the secret as a copilot-managed AWS SSM
parameter, however the flow execution containers are launched on-demand
by the ecs-worker service (according to the architechture described at
https://docs.prefect.io/integrations/prefect-aws/ecs_guide.

It's not clear to me how to make those secrets available as env vars in
the on-demand flow executor images. I don't see a way to leverage the
prefect deployment config yaml files to embed those secrets as env
variables (no apparent way to get that level of control over the
ephemeral task definitions the ecs-worker generates for the flow
execution images).

I assume it's possible somehow, but instead of going down that hole
I'm going to switch over to the obvious alternative - using prefect
cloud-stored secrets.