Enable adhoc script execution on ecs-worker container.
Use case: I wanted to register a baseline model on the prod mlflow
tracking server. I think I'll only do this now and then, and is a good
example of needing a way to run adhoc scripts that can use my core and
flows python modules. Running them on the ecs-worker is a quick way to
create a prod python environment that has proper permissions to
communicate with mlflow.

Also updated some scripts to make ssh-ing to different ECS tasks easier.