[WIP] Initialize chunked EIA data in DVC.
Switching to a chunked dataset system for the DVC data warehouse for
EIA demand data. This will allow us to rely more on DVC and less on
the live EIA API - which will save time on API requests and grant
access to data back to 2015 which the API recently stopped serving.
It will also save on storage, as the current strategy repeatedly adds
redundant data when ranges intersect.

The dataset chunks are 3-month (Quarterly) chunks of the EIA timeseries
data. An index has been created to reference the chunks and indicate
whether they are 'complete' (have data for every hour they logically cover).

This commit adds temporary logic to create the initial chunks and index.
Next step is to clean things up and update ETL to check the index, fetch
what isn't already covered by DVC chunks, and add the new data to the
chunks. After that, the training flow both training and eval flows will
rely entirely on DVC (including for backfill) and not at all on the EIA
API.