[WIP] First (buggy) pass at overwriting chunk index with weather.
Found a bug, but commiting the WIP version before I change approaches.
Currently, before overwiting the index I'm trying to take advantage of
the cached EIA data. But that's unecessarily complex (and already lead
to a few of the bugs I've found). Instead it would be better to just
blow away the cache at the beginning. This means fetching more data via
API on runs where we overwrite DVC - but that's ok, and its more correct
- when overwriting DVC one would resonably expect it would start from a
  clean slate rather than partially reuse the old DVC cache.